twitter_analytics/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
tweepy==3.8.0
psycopg2==2.8.4
django-filter==2.2.0
